move unit test back to integration test
