+++ libs/vertexai/tests/integration_tests/test_chat_models.py
@@ -6,6 +6,11 @@ from typing import List, Optional, cast
 
 import pytest
 from google.cloud import storage
+from google.cloud.aiplatform_v1beta1.types import (
+    Blob,
+    Content,
+    Part,
+)
 from langchain import agents
 from langchain_core.messages import (
     AIMessage,
@@ -27,6 +32,9 @@ from langchain_google_vertexai import (
     HarmBlockThreshold,
     HarmCategory,
     create_context_cache,
+)
+from langchain_google_vertexai.chat_models import (
+    _parse_chat_history_gemini,
 )
 from tests.integration_tests.conftest import _DEFAULT_MODEL_NAME
 
@@ -262,14 +270,14 @@ def test_multimodal_media_inline_base64_template(file_uri, mime_type) -> None:
     assert isinstance(output.content, str)
 
 
-@tool
-def get_pixel_info(query: str):
-    """Retrieves information about the Pixel."""
-    return "MOCK PIXEL INFO STRING"
-
-
-@pytest.mark.release
+@pytest.mark.extended
 def test_multimodal_media_inline_base64_agent() -> None:
+
+    @tool
+    def get_climate_info(query: str):
+        """Retrieves information about the Climate."""
+        return "MOCK CLIMATE INFO STRING"
+
     llm = ChatVertexAI(model_name="gemini-1.5-pro-001")
     prompt_template = ChatPromptTemplate.from_messages(
         [
@@ -293,13 +301,42 @@ def test_multimodal_media_inline_base64_agent() -> None:
     }
     text_message = {"type": "text", "text": "Describe the attached media in 5 words."}
     message = [media_message, text_message]
-    tools = [get_pixel_info]
-    agent = agents.create_tool_calling_agent(llm, tools, prompt_template)
+    tools = [get_climate_info]
+    agent = agents.create_tool_calling_agent(
+        llm=llm, tools=tools, prompt=prompt_template)
     agent_executor = agents.AgentExecutor(
         agent=agent, tools=tools, verbose=False, stream_runnable=False
     )
     output = agent_executor.invoke({"input": message})
-    assert isinstance(output, str)
+    assert isinstance(output["output"], str)
+
+
+def test_parse_history_gemini_multimodal_FC():
+    storage_client = storage.Client()
+    # Can't use the pixel.mp3, since it has too many tokens it will hit quota
+    # error.
+    file_uri = (
+        "gs://cloud-samples-data/generative-ai/audio/audio_summary_clean_energy.mp3"
+    )
+    mime_type = "audio/mp3"
+    blob = storage.Blob.from_string(file_uri, client=storage_client)
+    media_base64 = base64.b64encode(blob.download_as_bytes()).decode()
+    media_message = {
+        "type": "media",
+        "data": media_base64,
+        "mime_type": mime_type,
+    }
+    instruction = "Describe the attached media in 5 words."
+    text_message = {"type": "text", "text": instruction}
+    message = str([media_message, text_message])
+    history = [HumanMessage(content=message)]
+    parts = [
+        Part(inline_data=Blob(data=media_base64, mime_type=mime_type)),
+        Part(text=instruction),
+    ]
+    expected = [Content(role="user", parts=parts)]
+    _, response = _parse_chat_history_gemini(history=history)
+    assert expected == response
 
 
 @pytest.mark.xfail(reason="investigating")
